+++ packages/mathlifier/src/lib/mathlifiers/djot.ts
@@ -1,20 +1,21 @@
-import { mathlifierTex } from "./tex";
+import { md } from "./md";
 
 /**
- * generates djot markup.
+ * generates djot markup by converting $x$ and $$x$$ to $`x` and $$`x`.
+ * ams math environments are put in display math mode
  *
  * math: starts with ${x}, terminates with new line. MathlifierDj will add $`x` delimiters.
  * display: starts with $${x}, terminates with empty line. MathlifierDj will add $$`x` delimiters.
- * ams env: starts with $${'align'}x, etc, terminates with empty line. MathlifierDj will add $$`\begin{env}x\end{env}` delimiters.
+ * ams env: starts with #${'align'}x, etc, terminates with empty line. MathlifierDj will add $$`\begin{env}x\end{env}` delimiters.
  * text: starts with @${x}, terminates immediately. MathlifierDj will interpolate these as regular strings
  *
  * mathlifier will also turn any non-escaped $x$ delimiters into djot math $`x` syntax.
  */
-export function mathlifierDj(
+export function dj(
   strings: TemplateStringsArray,
   ...values: unknown[]
 ): string {
-  const markup = mathlifierTex(strings, ...values);
+  const markup = md(strings, ...values);
   return markup
     .replace(/(?<!\\)\$(?!`)\$([^]+?)\$\$/g, (_, match) => `$$\`${match}\``)
     .replace(